# server/mcp/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "evcc",
    "description": "Solar charging. Super simple.",
    "version": "0.2.0",
    "contact": {
      "url": "https://github.com/evcc-io/evcc"
    }
  },
...
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "Login",
        "description": "Administrator login. Returns authorization cookie required for all protected endpoints.",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "$ref": "#/components/schemas/Password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Set-Cookie": {
                "description": "The authorization cookie",
                "schema": {
                  "type": "string",
                  "example": "auth=ey...qY; Path=/; Expires=Fri, 04 Apr 2025 14:35:37 GMT; HttpOnly; SameSite=Strict"
                }
              }
            }
          },
          "401": {
            "description": "Invalid password provided"
          }
        }
      }
    },
...
    "/auth/password": {
      "put": {
        "operationId": "changePassword",
        "summary": "Change admin password",
        "description": "Changes the admin password.",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePassword"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid password provided"
          }
        }
      }
    },
...
    "/auth/apikey": {
      "get": {
        "operationId": "getApiKeyStatus",
        "summary": "API key status",
        "description": "Reports whether an API key has been generated. The key itself is\nonly returned once at creation time (POST), never on subsequent\nreads.\n",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "configured": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "operationId": "regenerateApiKey",
        "summary": "Generate or rotate the API key",
        "description": "Generates a fresh API key, replacing any existing one. The\nreturned key is shown only once; store it immediately.\n\nEven when the request is authenticated via API key (Bearer), the\nadmin password must be supplied in the request body to prevent a\nleaked key from rotating itself. The password check is skipped\nwhen the server is started with `--disable-auth`.\n",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "$ref": "#/components/schemas/Password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string",
                      "description": "The new API key (cleartext, shown once).",
                      "example": "evcc_aB3xYz7k0Pq2sN5mWtR4"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid admin password"
          },
          "403": {
            "description": "Forbidden in demo mode"
          }
        }
      }
    },
    "/batterydischargecontrol/{enable}": {
      "post": {
        "operationId": "setBatteryDischargeControl",
        "summary": "Control battery-discharge",
        "description": "Prevent home battery discharge during vehicle fast charging.",
        "externalDocs": {
          "url": "https://docs.evcc.io/en/features/battery"
        },
        "tags": [
          "battery"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/enable"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/BooleanResult"
          }
        }
      }
    },
    "/batterygriddischarge/{enable}": {
      "post": {
        "operationId": "setBatteryGridDischarge",
        "summary": "Control battery grid discharge",
        "description": "Allow the home battery to discharge to the grid (experimental).",
        "externalDocs": {
          "url": "https://docs.evcc.io/en/features/battery"
        },
        "tags": [
          "battery"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/enable"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/BooleanResult"
          }
        }
      }
    },
...
    "/batterygridchargelimit/{cost}": {
      "post": {
        "operationId": "setBatteryGridChargeLimit",
        "summary": "Set battery grid charge limit",
        "description": "Charge home battery from grid when price or emissions are below the threshold. Uses price if a dynamic tariff exists. Uses emissions if a CO₂-tariff is configured. Ignored otherwise.",
        "externalDocs": {
          "url": "https://docs.evcc.io/en/features/battery"
        },
        "tags": [
          "battery"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/costLimit"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/NumberResult"
          }
        }
      }
    },
...
    "/buffersoc/{soc}": {
      "post": {
        "operationId": "setBufferSoc",
        "summary": "Set battery buffer SoC",
        "description": "Set battery buffer SoC.",
        "externalDocs": {
          "url": "https://docs.evcc.io/en/features/battery"
        },
        "tags": [
          "battery"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/soc"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/NumberResult"
          }
        }
      }
    },
    "/bufferstartsoc/{soc}": {
      "post": {
        "operationId": "setBufferStartSoc",
        "summary": "Set battery buffer start SoC",
        "description": "Set battery buffer start SoC.",
        "externalDocs": {
          "url": "https://docs.evcc.io/en/features/battery"
        },
        "tags": [
          "battery"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/soc"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/NumberResult"
          }
        }
      }
    },
...
    "/prioritysoc/{soc}": {
      "post": {
        "operationId": "setPrioritySoc",
        "summary": "Set battery priority SoC",
        "description": "Set battery priority SoC.",
        "externalDocs": {
          "url": "https://docs.evcc.io/en/features/battery"
        },
        "tags": [
          "battery"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/soc"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/NumberResult"
          }
        }
      }
    },
    "/residualpower/{power}": {
      "post": {
        "operationId": "setResidualPower",
        "summary": "Set grid residual power",
        "description": "Set grid connection operating point.",
        "externalDocs": {
          "url": "https://docs.evcc.io/en/reference/configuration/site#residualpower"
        },
        "tags": [
          "battery"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/power"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/NumberResult"
          }
        }
      }
    },
    "/gridexportlimit/{power}": {
      "post": {
        "operationId": "setGridExportLimit",
        "summary": "Set grid export limit",
        "description": "Set the static grid export power limit in W used as optimizer constraint. 0 disables the limit. An active HEMS curtailment takes precedence.",
        "tags": [
          "general"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/power"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/NumberResult"
          }
        }
      }
    },
    "/solaradjusted/{enable}": {
      "post": {
        "operationId": "setSolarAdjusted",
        "summary": "Adjust solar forecast",
        "description": "Adjust the solar forecast to real production data of the current day.",
        "tags": [
          "experimental"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/enable"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/BooleanResult"
          }
        }
      }
    },
    "/config/solar-min-soc": {
      "get": {
        "operationId": "getSolarMinSocConfig",
        "summary": "Get forecast-based vehicle minimum charge configuration",
        "tags": [
          "general"
        ],
        "responses": {
          "200": {
            "description": "Configuration and current forecast state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolarMinSocStatus"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "setSolarMinSocConfig",
        "summary": "Set forecast-based vehicle minimum charge configuration",
        "tags": [
          "general"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolarMinSocConfig"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated configuration and current forecast state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolarMinSocStatus"
                }
              }
            }
          },
          "400": {
            "description": "Invalid thresholds, vehicle name, or minimum charge value"
          }
        }
      }
    },
...
    "/system/log/areas": {
      "get": {
        "operationId": "getLogAreas",
        "summary": "List of all log areas",
        "description": "Returns a list of all log areas (e.g. `lp-1`, `site`, `db`).",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogAreas"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangePassword": {
        "type": "object",
        "properties": {
          "current": {
            "$ref": "#/components/schemas/Password"
          },
          "new": {
            "$ref": "#/components/schemas/Password"
          }
        }
      },
...
      "Password": {
        "description": "Admin password",
        "type": "string"
      },
...
    },
    "parameters": {
...
      "enable": {
        "name": "enable",
        "description": "Charging mode.",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "enum": [
            "true",
            "false",
            "1",
            "0"
          ]
        }
      }
    },
    "responses": {
...
      "BooleanResult": {
        "description": "Success - Boolean result",
        "content": {
          "application/json": {
            "schema": {
              "type": "boolean"
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Unauthorized - Login and try again",
        "content": {
          "text/plain": {
            "schema": {
              "type": "string",
              "enum": [
                "Unauthorized"
              ]
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Long-lived API key (prefixed `evcc_`, managed via /auth/apikey).\nGrants access to /api/db/* without requiring the X-Admin-Password header.\n"
      }
    }
  }
}